+++ src/test/java/com/activecourses/upwork/UpworkApplicationTests.java
@@ -1,10 +1,10 @@
-package com.upwork.clone.backend;
+package com.activecourses.upwork;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class BackendApplicationTests {
+class UpworkApplicationTests {
 
 	@Test
 	void contextLoads() {
